requirements.txt
flask
flask_sqlalchemy
redis
mysqlclient
requests
gunicorn
bs4